refactor: extract figure-manager.js, move buildSummaryPage to summary-generate.js
- figure-manager.js: 373 lines — openPanel(), getDataThenPlot(),
  cleanPlot(), cleanPanel() as public API; createEqPanel1D() and
  registerButtons() as internals. Orchestrates all plot-type-specific
  behaviour.
- summary-generate.js: buildSummaryPage(openPanel) now lives alongside
  generateSummary(). Receives openPanel as callback for jump buttons.
- index.js: 430 → 113 lines — pure bootstrap; imports and load
  handler wiring only. Final state: 74% reduction from original 1316.
