+++ public/javascripts/summary-generate.js
@@ -1,4 +1,6 @@
-import { getStatusBar } from './status-bar.js';
+import { getStatusBar, wrap, addLoadingIndicator } from './status-bar.js';
+import { getBasicParameters } from './units.js';
+import { requestPlotData } from './api.js';
 
 export async function generateSummary(data) {
     const container = document.querySelector('#container');
@@ -519,4 +521,42 @@ function minmax(as, padding = 0) {
         max = Math.max(max, v);
     });
     return [min - (max - min) * padding, max + (max - min) * padding];
+}
+
+// ------------------------------------------------------------------
+//  Page building (imported by figure-manager)
+// ------------------------------------------------------------------
+
+/**
+ * Build the summary page from equilibrium data and register jump
+ * buttons that switch to the corresponding plot panels.
+ *
+ * @param {Function} openPanel — The tab-panel opener (bound to a button).
+ */
+export async function buildSummaryPage(openPanel) {
+    await getBasicParameters();
+    const summary_data = await (await requestPlotData('Summary')).json();
+    const summaryContainer = await generateSummary(summary_data);
+
+    if (summaryContainer === undefined) {
+        // summary page is already generated
+        return;
+    }
+
+    // register jump button on summary page
+    summaryContainer.querySelectorAll('.summary-jump-button').forEach(btn => {
+        btn.addEventListener(
+            'click',
+            wrap(async e => {
+                e.preventDefault();
+                const panelSwitch = document.querySelector(
+                    `#${btn.id.split('-')[1]}`
+                );
+                await addLoadingIndicator(openPanel.bind(panelSwitch))();
+
+                panelSwitch.checked = true;
+                document.querySelector(`#${btn.id.slice(8)}`).click();
+            })
+        );
+    });
 }
